// src/types/resources/MedicationRequest.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TDateTime } from '../simpleTypes/DateTime';
import { TCanonical } from '../simpleTypes/Canonical';
import { TAnnotation } from '../partials/Annotation';
import { TDosage } from '../partials/Dosage';
import { TMedicationRequestDispenseRequest } from '../partials/MedicationRequestDispenseRequest';
import { TMedicationRequestSubstitution } from '../partials/MedicationRequestSubstitution';

export type TMedicationRequest = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  status: string;
  statusReason?: TCodeableConcept;
  intent: string;
  category?: TCodeableConcept[];
  priority?: string;
  doNotPerform?: boolean;
  reportedBoolean?: boolean;
  reportedReference?: TReference;
  medicationCodeableConcept?: TCodeableConcept;
  medicationReference?: TReference;
  subject: TReference;
  encounter?: TReference;
  supportingInformation?: TReference[];
  authoredOn?: TDateTime;
  requester?: TReference;
  performer?: TReference;
  performerType?: TCodeableConcept;
  recorder?: TReference;
  reasonCode?: TCodeableConcept[];
  reasonReference?: TReference[];
  instantiatesCanonical?: TCanonical[];
  instantiatesUri?: TUri[];
  basedOn?: TReference[];
  groupIdentifier?: TIdentifier;
  courseOfTherapyType?: TCodeableConcept;
  insurance?: TReference[];
  note?: TAnnotation[];
  dosageInstruction?: TDosage[];
  dispenseRequest?: TMedicationRequestDispenseRequest;
  substitution?: TMedicationRequestSubstitution;
  priorPrescription?: TReference;
  detectedIssue?: TReference[];
  eventHistory?: TReference[];
};